+++ src/Support/Changelog.md
@@ -10,6 +10,8 @@
 - Added PHP 8.3 Support
 - Changed minimum Laravel version to v10.38.2
 - Added the `Features` accessor class, which is a syntactic sugar for areas of the configuration
+- Added the `Merchant` DTO class, that complies with the common, v4 `Merchant` interface
+- Added the `Address` DTO class (implements the `Address` interface)
 
 ## 3.x Series